PPM Performance pie chart

+++ MaintenanceWebApp/Data/PPMStatusByYear.cs
@@ -6,7 +6,7 @@ namespace MaintenanceWebApp.Data
     public class PPMStatusByYear
     {
         [Key]
-        public int RowNum { get; set; }
+        public long RowNum { get; set; }
         public int Tasks { get; set; }
         public int Year { get; set; }
         public string Status { get; set; }
